+++ src/executor.rs
@@ -12,10 +12,10 @@ use tracing::{debug, instrument, warn};
 const EXECUTOR_CONTROL_CHANNEL_SIZE: usize = 1024;
 
 pub(crate) trait JobExecutor {
+    async fn work(&self) -> Result<JobId>;
     async fn enqueue(&self, job: Job) -> Result<JobId>;
     async fn cancel(&self, id: &JobId) -> Result<()>;
     async fn state(&self, id: &JobId) -> Result<JobState>;
-    async fn work(&self) -> Result<JobId>;
     async fn shutdown(self, opts: ShutdownOpts) -> Result<()>;
 }
 
@@ -115,9 +115,23 @@ impl JobExecutor for Executor {
                             self.jobs.write().await.state.insert(id.clone(), JobState::Running);
                             Ok(id)
                         },
-                        ChangeExecutorStateEvent::JobCancelled(id) => {
+                        ChangeExecutorStateEvent::JobCanceled(id) => {
                             {
-                                self.jobs.write().await.state.insert(id.clone(), JobState::Cancelled);
+                                self.jobs.write().await.state.insert(id.clone(), JobState::Canceled);
+                            }
+                            self.requeue_jobs().await?;
+                            Ok(id)
+                        },
+                        ChangeExecutorStateEvent::JobTimeout(id) => {
+                            {
+                                self.jobs.write().await.state.insert(id.clone(), JobState::Timeout);
+                            }
+                            self.requeue_jobs().await?;
+                            Ok(id)
+                        },
+                        ChangeExecutorStateEvent::JobError(id) => {
+                            {
+                                self.jobs.write().await.state.insert(id.clone(), JobState::Error);
                             }
                             self.requeue_jobs().await?;
                             Ok(id)
@@ -187,7 +201,9 @@ impl JobExecutor for Executor {
 #[allow(clippy::enum_variant_names)]
 pub(crate) enum ChangeExecutorStateEvent {
     JobStarted(JobId),
-    JobCancelled(JobId),
+    JobCanceled(JobId),
+    JobTimeout(JobId),
+    JobError(JobId),
     JobCompleted(JobId),
 }
 
@@ -210,7 +226,7 @@ mod test {
         });
         let job_0 = task_0.job.clone();
         let job_id_0 = JobId::new("task 0 id");
-        let job_0 = Job::new(job_id_0.clone(), job_0);
+        let job_0 = Job::new(job_id_0.clone(), job_0, None);
 
         let task_1 = Task::new(TaskSchedule::Once, |_id| {
             Box::pin(async move {
@@ -219,7 +235,7 @@ mod test {
         });
         let job_1 = task_1.job.clone();
         let job_id_1 = JobId::new("task 1 id");
-        let job_1 = Job::new(job_id_1.clone(), job_1);
+        let job_1 = Job::new(job_id_1.clone(), job_1, None);
 
         executor.enqueue(job_0).await.unwrap();
         executor.enqueue(job_1).await.unwrap();
@@ -258,10 +274,7 @@ mod test {
             // 1st - Canceled, 2nd - Running
             async {
                 tokio::time::sleep(Duration::from_millis(600)).await;
-                assert_eq!(
-                    executor.state(&job_id_0).await.unwrap(),
-                    JobState::Cancelled
-                );
+                assert_eq!(executor.state(&job_id_0).await.unwrap(), JobState::Canceled);
             },
             async {
                 tokio::time::sleep(Duration::from_millis(600)).await;