+++ src/worker.rs
@@ -28,10 +28,16 @@ enum ChangeStateEvent {
     Shutdown(ShutdownOpts),
 }
 
-pub struct Worker {
+pub(crate) struct Worker {
     tokio_handler: Option<JoinHandle<()>>,
     thread_handler: Option<std::thread::JoinHandle<()>>,
     channel: Sender<ChangeStateEvent>,
+}
+
+#[derive(Debug)]
+enum JobExecutionResult {
+    Completed,
+    Timeout,
 }
 
 impl Worker {
@@ -91,12 +97,13 @@ impl Worker {
         worker_channel: ControlChannel<ChangeStateEvent>,
     ) {
         let mut ids: Vec<JobId> = Vec::new();
-        let mut handlers: Vec<JoinHandle<()>> = Vec::new();
+        let mut handlers: Vec<JoinHandle<JobExecutionResult>> = Vec::new();
 
         // Push a single always-pending job to avoid panics on empty select_all
         let fake_id = JobId::new(Uuid::new_v4());
-        let fake_handler =
-            tokio::task::spawn(Box::pin(async { futures::future::pending::<()>().await }));
+        let fake_handler = tokio::task::spawn(Box::pin(async {
+            futures::future::pending::<JobExecutionResult>().await
+        }));
         ids.push(fake_id);
         handlers.push(fake_handler);
 
@@ -110,11 +117,24 @@ impl Worker {
                             ChangeStateEvent::StartJob(job) => {
                                 let id = job.id();
                                 let job_to_run = &job.job();
-                                let mut job_to_run = job_to_run.write().await;
-                                let job_to_run = (job_to_run)(id);
-                                let handler = tokio::task::spawn(Box::pin(job_to_run));
+                                let job_to_run = (job_to_run.write().await)(id);
+
                                 ids.push(job.id());
-                                handlers.push(handler);
+
+                                if let Some(timeout) = job.timeout() {
+                                    let handler = tokio::task::spawn(Box::pin(async move {
+                                        select! {
+                                            _ = job_to_run => {JobExecutionResult::Completed},
+                                            _ = tokio::time::sleep(timeout) => {JobExecutionResult::Timeout}
+                                            }
+                                        }
+                                    ));
+                                    handlers.push(handler);
+                                } else {
+                                    let handler = tokio::task::spawn(Box::pin(async move {job_to_run.await; JobExecutionResult::Completed}));
+                                    handlers.push(handler);
+                                }
+
                                 let _ = executor_channel
                                     .send(ChangeExecutorStateEvent::JobStarted(job.id()))
                                     .await
@@ -129,7 +149,7 @@ impl Worker {
                                     let handler = handlers.remove(index);
                                     handler.abort();
                                     let _ = executor_channel
-                                        .send(ChangeExecutorStateEvent::JobCancelled(id))
+                                        .send(ChangeExecutorStateEvent::JobCanceled(id))
                                         .await
                                         .map_err(|_e| Error::SendingChangeStateEvent);
                                 }
@@ -170,17 +190,38 @@ impl Worker {
                     }
                 },
                 completed = select_all(&mut handlers.iter_mut()) => {
-                    let (_, index, _) = completed;
+                    let (result, index, _) = completed;
                     let id = ids.get(index);
-                    debug!(job_id = ?id, "job completed");
+
                     if let Some(id) = id {
                         let id = id.clone();
                         ids.remove(index);
                         handlers.remove(index);
+
+                        let job_completion_state = match result {
+                            Ok(completion_result) => {
+                                match completion_result {
+                                    JobExecutionResult::Completed => {
+                                        debug!(job_id = ?id, "job completed successfully");
+                                        ChangeExecutorStateEvent::JobCompleted(id)
+                                    },
+                                    JobExecutionResult::Timeout => {
+                                        debug!(job_id = ?id, "job killed because timeout");
+                                        ChangeExecutorStateEvent::JobTimeout(id)
+                                    },
+                                }
+                            },
+                            Err(error) => {
+                                debug!(job_id = ?id, %error, "job finished with error");
+                                ChangeExecutorStateEvent::JobError(id)
+                            },
+                        };
                         let _ = executor_channel
-                            .send(ChangeExecutorStateEvent::JobCompleted(id))
+                            .send(job_completion_state)
                             .await
                             .map_err(|_e| Error::SendingChangeStateEvent);
+                    } else {
+                        debug!(?result, "unexpected job finished")
                     }
                 }
             }